+++ src/main/java/com/buuz135/industrial/tile/misc/ResourcefulFurnaceTile.java
@@ -3,14 +3,17 @@ package com.buuz135.industrial.tile.misc;
 import com.buuz135.industrial.proxy.BlockRegistry;
 import com.buuz135.industrial.proxy.FluidsRegistry;
 import com.buuz135.industrial.tile.CustomElectricMachine;
+import com.buuz135.industrial.utils.ItemStackUtils;
 import com.buuz135.industrial.utils.Reference;
 import com.buuz135.industrial.utils.WorkUtils;
 import net.minecraft.item.EnumDyeColor;
+import net.minecraft.item.ItemStack;
 import net.minecraft.item.crafting.FurnaceRecipes;
 import net.minecraft.util.ResourceLocation;
 import net.minecraftforge.fluids.FluidStack;
 import net.minecraftforge.fluids.IFluidTank;
 import net.minecraftforge.items.IItemHandler;
+import net.minecraftforge.items.ItemStackHandler;
 import net.ndrei.teslacorelib.gui.BasicRenderedGuiPiece;
 import net.ndrei.teslacorelib.gui.BasicTeslaGuiContainer;
 import net.ndrei.teslacorelib.gui.IGuiContainerPiece;
@@ -36,7 +39,7 @@ public class ResourcefulFurnaceTile extends CustomElectricMachine {
         super.initializeInventories();
         input = addSimpleInventory(3, "input", EnumDyeColor.BLUE, "Input items", new BoundingRectangle(54, 25, 18, 18 * 3), (i, slot) -> !FurnaceRecipes.instance().getSmeltingResult(i).isEmpty(), (o, slot) -> false, false, null);
         output = addSimpleInventory(3, "Output", EnumDyeColor.ORANGE, "Output items", new BoundingRectangle(54 + 32 + 16, 25, 18, 18 * 3), (i, slot) -> !FurnaceRecipes.instance().getSmeltingResult(i).isEmpty(), (o, slot) -> false, false, null);
-        tank = addSimpleFluidTank(8000, "Tank", EnumDyeColor.LIME, 128, 25, FluidTankType.OUTPUT, (o) -> null, (o) -> false);
+        tank = addSimpleFluidTank(8000, "Tank", EnumDyeColor.LIME, 128, 25, FluidTankType.OUTPUT, (o) -> null, (o) -> true);
     }
 
     @Override
@@ -63,5 +66,15 @@ public class ResourcefulFurnaceTile extends CustomElectricMachine {
         pieces.add(new BasicRenderedGuiPiece(54 + 18 + 2, 26 + 18, 25, 18, new ResourceLocation(Reference.MOD_ID, "textures/gui/jei.png"), 24, 5));
         pieces.add(new BasicRenderedGuiPiece(54 + 18 + 2, 26 + 18 * 2, 25, 18, new ResourceLocation(Reference.MOD_ID, "textures/gui/jei.png"), 24, 5));
         return pieces;
+    }
+
+    @Override
+    protected boolean acceptsFluidItem(ItemStack stack) {
+        return ItemStackUtils.acceptsFluidItem(stack);
+    }
+
+    @Override
+    protected void processFluidItems(ItemStackHandler fluidItems) {
+        ItemStackUtils.processFluidItems(fluidItems, tank);
     }
 }